Twitter-Data/requirements.txt
tweepy

kafka-python